Switch meta queries to use a meta executor
This way meta queries can be planned separately and not rely on looking
at specific shards.

This is a bit ugly and we likely should refactor how the
IteratorCreator's are created to begin with. It's a bit fragile since we
have to handle some meta queries through shards and some through the
database index through two separate code paths. This will work in the
interim to continue enabling fast `SHOW MEASUREMENTS` and `SHOW TAG
VALUES` queries.